pages init + NotFound page
