+++ package.json
@@ -19,6 +19,7 @@
     "react-dom": "^18.2.0",
     "react-intersection-observer": "^9.4.1",
     "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.0",
     "redux-thunk": "^2.4.2",
