+++ src/app.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import AppHeader from './components/app-header';
+import ErrorBoundary from "./helpers/error-boundary";
+import {BrowserRouter, Route, Routes} from "react-router-dom";
+import Constructor from "./pages/constructor";
+import NotFound from "./pages/not-found";
+
+function App() {
+  return (
+    <ErrorBoundary>
+      <AppHeader/>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Constructor/>}/>
+          <Route path="*" element={<NotFound/>}/>
+        </Routes>
+      </BrowserRouter>
+    </ErrorBoundary>
+  );
+}
+
+export default App;
