+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './components/app/app';
+import App from './app';
 import reportWebVitals from './reportWebVitals';
 import {compose, createStore, applyMiddleware} from 'redux';
 import {Provider} from 'react-redux';
