+++ src/pages/constructor.module.css
@@ -1,4 +1,4 @@
-.app {
+.constructor {
   display: flex;
   justify-content: center;
   gap: 40px;
