+++ src/pages/not-found.module.css
@@ -1,0 +1,8 @@
+.container {
+  display: flex;
+  flex-direction: column;
+}
+
+.content {
+  margin: auto;
+}
